Refactor console (part 1)
